app/pyproject.toml
[tool.poetry]
name = "financial-agent-system"
version = "0.1.0"
description = "A financial management system with AI agents"
authors = ["erisonsuzuki"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
sqlalchemy = "^2.0.29"
psycopg = {extras = ["binary"], version = "^3.1.18"}
yfinance = "^0.2.40"
langchain = "^0.2.1"
langchain-google-genai = "^1.0.4"
langchain-community = "^0.2.1"
python-dotenv = "^1.0.1"
httpx = "^0.27.0"
pyyaml = "^6.0.1"
pandas = "^2.2.2"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
argon2-cffi = "^23.1.0"
email-validator = "^2.1.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
respx = "^0.21.0"
